added Memory efficient coding tip
I've noticed a couple times that freq.cfg can get extremely big compared to the actual data in a freq struct. E.g. a struct array of 16 freq's was 162 MB, but after doing 
[freq.cfg] = deal([]); 
it was only 29 MB! ft_analysispipeline showed me a large number of cfg's for each run. I guess the reason is that ft_appenddata etc keeps the cfg of each single inputfiles...