Update 2021-01-12-release.md
an important thing to note is that I would like for the FT home page the list of differences/commits on github to be relative to the previous homepage announcement, not relative to the previous silent release. In this case I changed `20210111` into `20201229`, since that is the latest publicly announced version. The changes in between that one and the current one are also relevant.   

A small thing is that links to reference documentation (aka the help) should point to github. There people can look up the code details and history. That change has been made everywhere on the FT website.