Update irasa.md
#IRASA (intro): The distinction between separation (Wen & Liu) and feature extraction (Stolk et al) is no longer prominent. This is the first section that readers encounter, and a good place for introducing the two-step structure of the example page (separation and extraction). Hence, I'd keep the original text here.

The simulated example also conducted 'feature extraction', as you will find in the script and the very first plot, so I don't really understand the 'distinction', however, I rephrased it, you can see if it's prominent enough now.

#extraction/localization: Following from the above, I'd suggest to keep my original edits here. Don't worry about "the original study", as people know what it refers to, and else they can simply click the link.

I'm still reluctant to say 'the original', sorry, instead I referred to Stolk et al. 2019 again.

#updates: This is the section that could benefit from some more editing because you introduce concepts that people may not be familiar with, e.g. "geometric/arithmetic means", "mtmfft", "90%", "1.9", "resampling factors". I'm afraid these things are clear only you and me but not to first-time readers wanting to just get a feel for how to use IRASA and how it could help them in extracting and localizing rhythmic spectral features. 

First, these basic concepts should be covered by users before they decide to use the method and there's a link to the original paper Wen and Liu 2016. Second, you had already introduce the algorithm of IRASA at the top of the page and explained the resampling, I agreed and didn't change anything. Finally, as explained the beginning of the update section, this section was made for those who used the old implementation and were directed to this page when the run into error of their matlab code, so they should know what cfg.method='mtmfft' is. For the 'first-time readers', they can use the example script directly, I don't a conflict of the contents here.

# So, either you introduce and describe those concepts in more detail, or you keep the updates section to a minimum in order to later add in your source IRASA pipeline. 

That's a different thing which is not in discussion of this PR.

#My suggestion would be the latter, and to briefly reference your github PR for the timing being as well as maybe the resampling section (though this needs some editing still to become more accessible to a general public).

I've explained why the update section was made and why a mere link of the PR of the new implementation is insufficient, hope you understand. The last paragraph is indeed not about the update per se, but I think it should be warned because careless using filtered data can lead to serious error in the results.